services/server/requirements.txt
Flask==1.1.2
Flask-Cors==3.0.10
flask-migrate==2.7.0
Flask-SQLAlchemy==2.5.1
gunicorn==20.1.0
psycopg2-binary==2.9.3
